hrl_sarp/requirements.txt
# ══════════════════════════════════════════════════════════════════════
# HRL-SARP: Hierarchical RL for Sector-Aware Risk-Adaptive Portfolio Mgmt
# Python 3.10+ required
# ══════════════════════════════════════════════════════════════════════

# ── Core Scientific Stack ────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0
scikit-learn==1.4.1

# ── Deep Learning ────────────────────────────────────────────────────
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1

# ── Graph Neural Networks ───────────────────────────────────────────
torch-geometric==2.5.0
torch-scatter==2.1.2
torch-sparse==0.6.18

# ── Reinforcement Learning Utilities ─────────────────────────────────
gymnasium==0.29.1

# ── NLP / Transformers ──────────────────────────────────────────────
transformers==4.38.2
tokenizers==0.15.2
datasets==2.18.0
sentencepiece==0.2.0

# ── Technical Analysis ──────────────────────────────────────────────
pandas-ta==0.3.14b1
ta-lib==0.4.28

# ── Financial Data ──────────────────────────────────────────────────
yfinance==0.2.36
kiteconnect==5.0.1
jugaad-data==0.23

# ── Hidden Markov Models ────────────────────────────────────────────
hmmlearn==0.3.0

# ── Explainability ──────────────────────────────────────────────────
shap==0.45.0
captum==0.7.0

# ── Visualization & Dashboard ───────────────────────────────────────
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0
streamlit==1.31.1

# ── Experiment Tracking ─────────────────────────────────────────────
mlflow==2.11.1

# ── Database ────────────────────────────────────────────────────────
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# ── Web Scraping ────────────────────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
lxml==5.1.0
selenium==4.18.1

# ── Configuration & Serialization ───────────────────────────────────
pyyaml==6.0.1
omegaconf==2.3.0
python-dotenv==1.0.1

# ── Scheduling / Pipeline ──────────────────────────────────────────
schedule==1.2.1
# apache-airflow==2.8.2  # Optional: uncomment for Airflow DAG support

# ── Report Generation ──────────────────────────────────────────────
jinja2==3.1.3
weasyprint==61.2

# ── Testing ─────────────────────────────────────────────────────────
pytest==8.0.2
pytest-cov==4.1.0

# ── Utilities ───────────────────────────────────────────────────────
tqdm==4.66.2
joblib==1.3.2
tenacity==8.2.3
pytz==2024.1
holidays==0.42
